Use binary encoding for "deduced" command arguments
Arguments with cpc.ptc=11 are "deduced" - meaning that their possible value depends on the some previous argument.

The PUS standard uses the word deduced in 3 contexts:
1. deduced presence - for example TC (11,6) where the presence or not of time tag 1 and 2 are based on the value selected for the type.
2. deduced size - like the mask in TC(12,5) whose size (but not other characteristics such as type or calibration) is determined by the selected parameter ID.
3. "classical" deduced - for example TC(20,3) where the "value" argument to be inserted is verified against the definition of the selected Parameter ID.

The SCOS2k ICD defines ptc = 11 without specifying which of the 3 is meant. Moreover none of the 3 is supported in the SCOS2k, and the SCOS2k MIB loader will skip these arguments with a warning.

This patch will create the arguments as binary with binary encoding.